Rearranging the code base.
